Update edmx to admintrainsql, only keeping new models that are news
Generation of the edmx was done differently than instructions because
instructions assume edmx is being synced to prod instead of train.
Connection settings were saved as CCTEntitiesTrain, connection string
has admintrainsql instead of adminprodsql, but everything else should be
the same as usual. Any models that were generated and new to the branch
were deleted besides StudentNews and StudentNews Category