Remove unnecessary and unused code
Following the pattern for many of the other endpoints, I had created repositories for the news
models. However, when I noticed that the use of these repositories did not appear to be
necessary for the news endpoints, I removed the news repositories. (See ActivityService.cs
line 63, or wherever the variable `record` is instantiated in GetActivitiesForSession(),
for an example of repository use.) Other code that many of the endpoints use despite how
few probably need it is the trimming code in the service methods. I commented this out so
that one last check can be made before it is deemed unnecessary.