Remove debug statement that was crashing people search
A select had been hard-coded into Advanced People Search that was only used for debugging purposes, but was running in production as well. It was searching for the building description of a specific student. That student has now graduated and so the query was returning nothing, which would cause the API to throw an error.