Refactor wellness to use unified table
This commit restructures the wellness service to make use of the Health Status table, atable that unifies standard checkins  and administrative overrides. This unified table can be accessed and updated through Entity Framework instead of needing additional stored procedures. The API now handles the expiration logic for standard check ins, and overrides are not marked differently except that they have a wider range of possible expiration dates and are created by a user instead of 360.WellnessCheck.