Simplify EventViewModel
The EventViewModel was previously using extension methods for XML null handling, but the C# null conditional operator does the same job more quickly and more readably.

Also, the EventViewModel is updated to only check for requirement 3 to find whether an event is public. Category 99 was to promote on a specific public calendar, but this is only a subset of public events.