Remove unneeded auth filter for GetEventsByTerm
The GetEventsByTerm route `api/chapel/:term` does not need additional StateYourBusiness auth filtering because it only gets data for the currently logged-in user. Since there was no case to handle a partial read of the ChapelEvent reource, the filter was causing all requests to fail with a 403 unauthorized status. The filter was previously in the EventService on the GetEventsForStudentByTerm method. However, the filter never gets called when it isn't in an ApiController class, which is why it was never revealed that the filter would always fail.