Update README for current VMs and CD
The README had outdated VM documentation. CS-360-API-TEST.gordon.edu no longer exists, and 360Train.gordon.edu is an alias for the 360-Frontend.gordon.edu server where the frontend is hosted. That machine shouldn't be a development machine in either case, since it's resources are needed for hosting the frontend, but also it has no connections to the backend, which is hosted on cts-360.gordon.edu.

Similarly, the Deployment instructions were outdated and misleading, since deployment of the API is currently manual.