Fix membership update issues
Because of a bug introduced a couple years ago, Memberships and membership privacies were failing to update for anyone except super admins. The API route for getting a membership object by membership ID was deleted, and it's usage in the StateYourBusiness auth filter was replaced with the GetMembershipsForStudent method, which returns an enumerable of memberships. The bug was in trying to cast this enumerable to a single Membership (instead of filtering). The fix was to use the GetSpecificMembership method which takes a membership id and returns that membership object. This alllows proper authorization so that these operations should now function as normal.