Add error logging to JobsController saveShiftForUser
when shifts are saved, some have started coming through with the Time part of DateTime lost, so the start and end time are both 12AM. This adds error logging to catch these errors, reject the improper shifts, and add a log to CCT's Error Log table detailing the erroneous shift.

This should allow users to access timesheets via 360 until we figure out the source of the issue, since people won't be able to submit erroneous shifts. It should also potentially help us track down the source of the issue.