Update custom authorizationFilter unauthorized message to fit default unauthorized message.
Our custom authorization filter (StateYourBusiness.cs) will return a `401 unauthorized` response when a user is not permitted to perform the specified operation on the specified resource. However, unlike the default unauthorized respone given when a requqest is unauthenticated, the custom response had no message body.

This was causing issues in our tests, so the custom unauthorized response has been updated to return the same message as the default unauthorized response.