Send unapproved image as default
Send unapproved image as default, in place of an approved ID image, when the user has no approved ID or preferred photo. This will never be shown to other users, since the unapproved image is only sent on the getMyImage route (which will return images only for the currently authenticated user).

Returning the unapproved image as "def" means that the front end will use it in the same way it would use an approved ID image. The unapproved image will never supercede approved or preferred images, since it only gets returned when we have no date on those.