Adjsut EventViewModel to represent one occurrence.
When we fetch events from 25Live's api (in XML format), every event has a list of one or more occurrences attached. An event represents a reservation that may be repeated multiple times in 25Live, whereas an Occurrence is a specific 'occurrence' of an event tied to a particular date, time, and location.

However, 360's model of events has always treated them as if the event was an occurrence tied to a particular datetime and location. The code before this achieved this by ignoring all but the first occurrence of an event. This change seeks to fix the conceptual gap. It does so by incorporating occurrence info into the EventViewModel directly, thereby shifting the definition of an event. In order to do this, we have one instance of a 25Live event *per occurrence*.

To accomplish this, I have adjusted the EventViewModel so that it stores data from only one occurrence, rather than a list of occurrences. We now construct our 360 events by by mapping every 25Live event with each of it's occurrences, resulting in a unique EventViewModel for each occurrence of a 25Live event.