Controller + Service
Navigating around issues with the backend. Still trying to understand what's going on all on my own. Dummy variables in place of where the actual SQL table headers are.

First commit to the original repo instead of my fork, after I was given the permissions to edit the original repo