Minimally working API with .NET 6
This is an initial commit of a minimally working configuration. At this point, the API starts up and responds to the route /api/activities with the full list of involvements from the database.

Authentication and authorization still definitely do not work. Until those are working, it's difficult to test other things. There's probably a lot left to do before this branch reaches parity with develop, but at least now we have a starting point.

As part of getting this working, I also upgraded from Entity Framework to Entity Framework Core. I'm not certain this was necessary, but I was unable to get the API working prior to the upgrade. The upgrade also comes with many benefits beyond the initial benefits of upgrading to .NET 6, and if we're going to have to test/refactor everything anyways, I figured we might as well do it all at once.