Working Azure AD Authentication
I have set the project up to use Azure AD for authentication, rather than local LDAP. This is in keeping with Gordon's other applications, and it also simplifies many things, includign project startup/configuration. It also opens the door to, in the future, using Azure AD to defines scopes/roles for resources rather than using StateYourBusiness and custom "Admin" tables to manage access rights.

The switch to Azure AD Authentication requires a significant reworking of the Auth logic in the UI. I have a prototype of these changes that is sufficient to login and access API resources that require authorization, but it is still in need of much polishing and - potentially - a deeper restructuring.