Save banners with correct image URL instead of local file path
When a banner is submitted to the API, it is stored in the API's local file path and the URL that it can be fetched from is stored in the database. However, we were storing the local filepath (e.g. `F:\\Sites\\360apitrain.gordon.edu\\browseable\\slider\\image.jpeg`) in the database instead of the web URL that the file can be retrieved from via HTTP (e.g `https://360apitrain.gordon.edu/browseable/slider/image.jpeg`).

Thus when the banners were loaded on the frontend, the browser would try to get the image from the local files instead of the web, which didn't work (both because the image didn't exist in the local file system and because browsers aren't allowed to access the local file system by default).

Now, we properly store the URL of the image in the database after uploading it to the local filesystem of the API.