make EventViewModel ids unique
Since we are now merging events with their occurrences, the original `event_id` id is no longer uinque. This was causing issues on the front end where only the first item with that id would be removed when filtering events.

Now, each EventViewModel has a unique Event_ID that is a combination of event id and occurrence id.