Allow alumni to login, fix auth issues
Previously, alumni have not been able to login. This was not an intentional design choice, but rather a bug relating to the organization of our Active Directory instance. While all student and facstaff accounts exist within the Gordon College organizational unit in AD, the Alumni OU does not. It is a sibling to the Gordon College OU. Our LDAP Authentication searched within the Gordon College OU for accounts, so it never found Alumni accounts.

This change addresses that issue by searching the entire Gordon domain for accounts. This has an unfortunate side effect that any user account within the gordon domain will be able to authenticate. However, because we require logged-in users to have a valid Gordon ID and profile info, this will exclude most accounts that we don't want using 360.

Furthermore, I have restructured how the `college_role` identity claim is assigned. This was because the alumni role was never assigned. Now, anyone in the `Alumni` OU will receive the alumni role.

Previously, `college_role` defaulted to FacStaff if no other special case was hit. That is very insecure. Instead, it now defaults to the `default` college_role, which is brand new and has no privileges at all. Since we use `college_role` for authorization in the API, this will prevent accounts from accessing data they shouldn't be able to.