Migrate CCT to standard DbContext structure
This migration also includes a refresh from the database. Since EF Core Power Tools (or maybe the underlying EF Core CLI) has been updated since the last refresh, it is able to more accurately detect which stored procedures produce a real output. In so doing, it has removed a LOT of empty boilerplate classes that used to represent a return value that didn't exist from the stored procedures.