Fix preBuild event preventing CI
The syntax for preBuild events in .csproj files has been changed. I have switched our preBuild event script (which puts the datetime and Git hash as of build time in a file for runtime reference) to the new syntax. The script is not working completely (I suspect an issue with preserving variables between commands), but it at least generates the needed version.txt file. Without the generated version.txt file, builds were failing especially in CI.