Fix Dining Service null balance
The Dining Service was returning a null balance for students because, after retrieving their balance, it was attempting to mutate the DiningInfoViewModel to update the CurrentBalance property, but this change wasn't affecting the collection returned by the Service. I have fixed the issue by making GetBalance a static method (which it already should've been since it doesn't depend on any instance methods), and calling it in the Select that produces the collection, thereby avoiding mutation entirely.