People Search: Exact match majors
We were previously searching majors with `MajorDescription.StartsWith(majorSearchTerm)`. This meant that, for example, a search for `Mathematics` majors would also show `Mathemtics Specialists` majors. Since we have an exhaustive dropdown of active majors in People Search, we should do exact matching so that people only see the majors they searched for.