Match chapel events to 25Live events properly and robustly
25Live event IDs were changed, but the join from chapel events was never updated, so the join always returned an empty array. I have updated the join to properly compare event IDs ignoring the concatenated Occurrence ID.

I also updated the process to be a Left Join, so that if a Chapel Event ever fails to match a 25Live event, it will still appear in the results (albeit lacking crucial data like name, description, location, and start/end date). This way, the event still appears in the users Attended Events list, even though we can't show them useful data about that event.