Update StateYourBusiness arguments and username comparison
StateYourBusiness previously did only unsafe casts when trying to fetch Arguments to a route. Now, it will instead perform a pattern match to determine if an argument exists as the right type.

Also, we were often comparing `username` (and some other strings) with the `==` operator, but this didn't account for different possible casings. Now, the comparison is case-insensitive in all cases.