Combine CI and CD workflows
Apparently, the `on: workflow_run` rule will always run on a repo's **default branch**, i.e. `develop` in our case. This means that we can't have a separate workflow for deployment that is configured to run after the CI workflow.

There are basically two solutions to this problem:
1. Combine CI and CD into a single workflow with several, dependent jobs.
2. Re-organize workflows so that they're only run when called, and create a new central workflow that will call the others when needed.

Option #1 is the simplest, and fits our needs best, so that is what I am doing.