Remove defunct site admin management
The Adminsitrator Service has become defunct because site admins are now managed via security group. I removed the unused controller, the last other usage of the service (which was broken because it used the database table that is no longer a source of truth for stie admin privileges) and the service and database mapping.